Ajout logo footer + fix indentation / balise
