+++ src/App.js
@@ -1,5 +1,5 @@
-import NavBar from './NavBar';
-import Footer from './Footer';
+import NavBar from './components/NavBar/NavBar';
+import Footer from './components/Footer/Footer';
 //import '../App.css';
 
 function App() {
