+++ src/components/Footer/footer.css
@@ -1,0 +1,4 @@
+footer {
+  display: flex;
+  justify-content: center;
+}
