--- src/styles/Footer.css
@@ -1,4 +1,0 @@
-.footer {
-    display: flex;
-    justify-content: center;
-}
