--- src/styles/NavBar.css
@@ -1,20 +1,0 @@
-a {
-    font-size: larger;
-    color: #FF6060;
-}
-
-.nav-bar {
-    display: flex;
-    justify-content: space-between;
-    flex-wrap: wrap;
-    margin: 3% 5%;
-}
-
-.main-link {
-    margin-right: 20px;
-}
-
-.header-img {
-    border-radius: 25px;
-    max-width: 90%;
-}
